requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Resume parsing
python-docx==1.2.0
pdfplumber==0.11.9
Pillow==12.1.0

# Embeddings (CPU only torch!)
sentence-transformers==5.2.2
torch --index-url https://download.pytorch.org/whl/cpu
